partie 1 finished

+++ ft_strlcat.c
@@ -6,19 +6,19 @@
 /*   By: lahammam <lahammam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 12:11:24 by lahammam          #+#    #+#             */
-/*   Updated: 2021/11/03 19:43:44 by lahammam         ###   ########.fr       */
+/*   Updated: 2021/11/06 14:14:02 by lahammam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, char *src, size_t size)
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	len_dest;
+	size_t			i;
+	size_t			len_dest;
 
 	i = 0;
-	len_dest = (unsigned int)ft_strlen(dest);
+	len_dest = ft_strlen(dest);
 	if (len_dest >= size)
 		return (ft_strlen(src) + size);
 	while (src[i] != '\0' && (len_dest + i < size - 1))
@@ -27,7 +27,7 @@ unsigned int	ft_strlcat(char *dest, char *src, size_t size)
 		i++;
 	}
 	dest[len_dest + i] = '\0';
-	return ((unsigned int)ft_strlen(src) + len_dest);
+	return (ft_strlen(src) + len_dest);
 }
 
 // int main()
